move files into app folder
