+++ src/renderer/index.tsx
@@ -2,12 +2,12 @@ import log from "electron-log";
 import React from "react";
 import { render } from "react-dom";
 
-import { createApp } from "./App";
+import { createApp } from "./app/create";
 import { installServices } from "./services/install";
 
 async function main() {
   const services = await installServices();
-  const App = createApp({ services });
+  const { App } = createApp({ services });
   render(
     <React.StrictMode>
       <App />
